Add helpers to build transaction for users (#50)
This adds "user transaction builder" types for loan and market contracts
providing functionality to build transactions to be signed and submitted
directly by a user.

The motivation is that the backend would nevertheless have to provide enough
information for the frontend to build the transactions itself. By also having
the backend doing this last step of constructing the transactions, we avoid the
need to embed contract ABIs in the frontend code.

Also sneak in some cleanups and improvements to tests.